requirements.txt
# Aplicación Soporte Administrativo - Dependencias

# Framework web
streamlit>=1.31.0

# Base de datos
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9

# Almacenamiento en la nube
cloudinary>=1.36.0

# IA y procesamiento
anthropic>=0.18.1

# Procesamiento de PDFs
pypdf>=4.0.1
reportlab>=4.0.9

# Procesamiento de Word
python-docx>=1.1.0

# Utilidades
python-dotenv>=1.0.1
